feat: Adds filtering of erc20 tokens by their l2 address (#447)
# What ❔

Sometimes we only want a specific set of tokens or even no tokens at all
to be finalized. If we want no erc20 tokens to be finalized, say, in the
event that we only want to do ETH withdrawals, this option can be set to

```
ONLY_FINALIZE_THESE_TOKENS="[]"
```

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `cargo fmt`.